xspn/requirements.txt
spflow
pycapnp
pytest